Added new features and fixes
Limited the sprites and bind it on the window sizes. Also the sprites
will now fade out after falling down and be removed from the DOM to
improve performance.

I also blocked multiple calls because it will turn down the performance
when you call to many sprites at once. now you have to wait for the
animation to finish.

- plus: I added a contribution-area :) have fun!